feat(backend): remove UserProfile model
Removing UserProfile model.  This is the first of several commits to
completely remove the model.  It deletes the admins, serializers, and
views, signals, etc.  associated w/ profiles.  It also removes the 1to1
relationship between profiles and users.  A future commit will remove
the model itself and replace it w/ a JSONField.