feat(backend): remove UserProfile model
This is the third of several commits to completely remove the model.  It
replaces the model w/ a JSONField.  That field can be synchronised
to/from the gateway.  I still need to write the actual synchronisation
function, but that must wait until after the authentication mechanism is
refactored and there is a gateway client to use.